requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.25.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-go>=0.20.0
tree-sitter-rust>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-c>=0.20.0
sentence-transformers>=2.2.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
anthropic>=0.18.0
openai>=1.0.0
boto3>=1.34.0
pymongo>=4.6.0
numpy>=1.24.0
arq>=0.25.0
redis>=5.0.0
neo4j>=5.0.0
networkx>=3.0
umap-learn>=0.5.0
hdbscan>=0.8.33
